Null ref error fixed for missing types
If the fully qualified type name was changed or removed from a project then the string would still be serialized but the inspector would throw a null-ref and not allow editing of the type. The value is now reset to the default 'None' value when this happens.
TODO: Should really have a 'Missing <typename>' display instead to aid in debugging.